Move out edit buttons to WorkSheetOperationsButtons,js file
